Reorganized the interface of the whole library
