+++ Benchmarks/Benchmarks/EncodeBase83Benchmark.cs
@@ -1,17 +1,17 @@
 using BenchmarkDotNet.Attributes;
-using Blurhash.Core;
+using Blurhash;
 
-namespace Benchmarks;
+namespace Benchmarks.Benchmarks;
 
 [MemoryDiagnoser]
 public class EncodeBase83Benchmark
 {
-    const int x = 5;
+    const int X = 5;
 
     [Benchmark(Baseline = true)]
     public void Original()
     {
         Span<char> output = stackalloc char[1];
-        x.EncodeBase83(output);
+        X.EncodeBase83(output);
     }
 }
