+++ Benchmarks/Benchmarks/FullStackDecodeBenchmark.cs
@@ -1,20 +1,17 @@
 using System.Drawing;
-using System.Drawing.Blurhash.DotNetFramework.Test;
-using System.Drawing.Common.Blurhash;
+using System.Drawing.Blurhash;
 using BenchmarkDotNet.Attributes;
 
-namespace Benchmarks;
+namespace Benchmarks.Benchmarks;
 
 [MemoryDiagnoser]
 public class FullStackDecodeBenchmark
 {
-    public static readonly Image SampleImage = Image.FromFile("Samples\\flower.jpg");
-
+    private static readonly Image SampleImage = Image.FromFile("Samples\\flower.jpg");
 
     [Benchmark(Baseline = true)]
     public void Original()
     {
-        var decoder = new Decoder();
-        decoder.Decode(ImageConversionTestCases.ExpectedHashes[(9, 9)], SampleImage.Width, SampleImage.Height).Dispose();
+        Blurhasher.Decode(ImageConversionTestCases.ExpectedHashes[(9, 9)], SampleImage.Width, SampleImage.Height).Dispose();
     }
 }
