+++ Blurhash.Core.Test/Blurhash.Core.Test.csproj
@@ -3,6 +3,7 @@
     <PropertyGroup>
         <TargetFramework>net5.0</TargetFramework>
         <IsPackable>false</IsPackable>
+        <RootNamespace>Blurhash.Tests</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
