+++ Blurhash.Core/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 ﻿using System.Runtime.CompilerServices;
 
 [assembly:InternalsVisibleTo("Blurhash.Core.Test")]
+[assembly:InternalsVisibleTo("Benchmarks")]
