+++ Blurhash.Core/Base83.cs
@@ -1,12 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Blurhash.Core
+namespace Blurhash
 {
     /// <summary>
     /// Contains methods to encode or decode integers to Base83-Strings
     /// </summary>
-    public static class Base83
+    internal static class Base83
     {
         internal const string Charset = @"0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz#$%*+,-.:;=?@[]^_{|}~";
