+++ Blurhash.Core/Blurhash.Core.csproj
@@ -3,8 +3,8 @@
   <PropertyGroup>
     <TargetFramework>netstandard1.2</TargetFramework>
     
-    <PackageId>Blurhash.Core</PackageId>
-    <PackageVersion>1.3.0</PackageVersion>
+    <PackageId>Blurhash</PackageId>
+    <PackageVersion>2.0.0</PackageVersion>
     <Authors>Markus Palcer</Authors>
     <Description>BlurHash is a compact representation of a placeholder for an image.</Description>
     <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
@@ -12,6 +12,8 @@
     <PackageTags>blurhash image preview compression encoding</PackageTags>
     <PackageProjectUrl>https://github.com/MarkusPalcer/blurhash.net</PackageProjectUrl>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
+
+    <RootNamespace>Blurhash</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
