+++ Blurhash.Microsoft.Extensions.System.Drawing/IBlurhasher.cs
@@ -1,0 +1,13 @@
+using System.Drawing;
+using Blurhash.Microsoft.Extensions.Core;
+
+namespace Blurhash.Microsoft.Extensions.System.Drawing
+{
+
+    /// <summary>
+    /// The blurhash algorithm abstraction for <see cref="Image"/>
+    /// </summary>
+    public interface IBlurhasher : IBlurhasher<Image>
+    {
+    }
+}
