+++ Blurhash.System.Drawing.Common/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 ﻿using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Blurhash.System.Drawing.Common.Test")]
+[assembly: InternalsVisibleTo("Benchmarks")]
